refactor(core): refinining modular structure
