+++ internal/modules/auth/services/service_test.go
@@ -11,9 +11,9 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/vitalfit/api/config"
-	authdomain "github.com/vitalfit/api/internal/auth/domain"
-	authmocks "github.com/vitalfit/api/internal/auth/mocks"
-	authservices "github.com/vitalfit/api/internal/auth/services"
+	authdomain "github.com/vitalfit/api/internal/modules/auth/domain"
+	authmocks "github.com/vitalfit/api/internal/modules/auth/mocks"
+	authservices "github.com/vitalfit/api/internal/modules/auth/services"
 	shared_errors "github.com/vitalfit/api/internal/shared/errors"
 	"github.com/vitalfit/api/internal/store"
 	mailermocks "github.com/vitalfit/api/pkg/mailer/mocks"
