+++ internal/shared/middleware/auth/auth.go
@@ -9,7 +9,7 @@ import (
 	"github.com/golang-jwt/jwt/v5"
 	"github.com/google/uuid"
 	appservices "github.com/vitalfit/api/internal/app/services"
-	authdomain "github.com/vitalfit/api/internal/auth/domain"
+	authdomain "github.com/vitalfit/api/internal/modules/auth/domain"
 )
 
 type AuthMiddleware struct {
