+++ internal/store/mocks.go
@@ -1,6 +1,6 @@
 package store
 
-import authmocks "github.com/vitalfit/api/internal/auth/mocks"
+import authmocks "github.com/vitalfit/api/internal/modules/auth/mocks"
 
 func NewMockStore() Storage {
 	return Storage{
